# backend/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Karat Hive API",
    "version": "0.1.0",
    "description": "Generated from Nest controllers + ZodValidationPipe schemas (`npm run openapi:generate`). Zod remains the single source of request shape (NFR-030, AD-BE-14). First artefact is partial: paths/methods and request Zod schemas are covered; response presenters are not yet dual-exported as Zod. Pre-code catalogue: docs/API-Route-Inventory.md — clients should treat this file as the evolving contract."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local API"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Karat Hive access JWT (not a Firebase ID token)."
      }
    },
    "schemas": {
      "SuccessMeta": {
        "type": "object",
        "properties": {
          "requestId": {
            "type": "string"
          },
          "serverTime": {
            "type": "string"
          },
          "nextCursor": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "requestId",
          "serverTime",
          "nextCursor"
        ]
      },
      "SuccessEnvelope": {
        "type": "object",
        "properties": {
          "data": {
            "nullable": true
          },
          "meta": {
            "$ref": "#/components/schemas/SuccessMeta"
          }
        },
        "required": [
          "meta"
        ]
      },
      "ErrorBody": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "nullable": true
          },
          "requestId": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ]
      }
    },
...
  },
  "paths": {
...
    "/v1/me/vendor": {
      "get": {
        "operationId": "VendorProfile_me",
        "tags": [
          "VendorProfile"
        ],
        "summary": "GET /v1/me/vendor",
        "description": "Handler VendorProfile.me. Request shapes come from ZodValidationPipe schemas (single source). Response DTOs are not yet Zod-exported — SuccessEnvelope is a placeholder.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success (envelope per Architecture §13 / EnvelopeInterceptor)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "VendorProfile_patch",
        "tags": [
          "VendorProfile"
        ],
        "summary": "PATCH /v1/me/vendor",
        "description": "Handler VendorProfile.patch. Request shapes come from ZodValidationPipe schemas (single source). Response DTOs are not yet Zod-exported — SuccessEnvelope is a placeholder.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "tradingName": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "description": {
                    "type": "string",
                    "maxLength": 2000
                  },
                  "contactPersonName": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100
                  },
                  "businessEmail": {
                    "type": "string",
                    "maxLength": 255,
                    "format": "email"
                  },
                  "legalBusinessName": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "tradeLicenceNumber": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 50
                  },
                  "businessAddress": {
                    "type": "string",
                    "minLength": 1
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success (envelope per Architecture §13 / EnvelopeInterceptor)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/me/vendor/performance/export": {
      "get": {
        "operationId": "VendorProfile_exportPerformance",
        "tags": [
          "VendorProfile"
        ],
        "summary": "GET /v1/me/vendor/performance/export",
        "description": "Handler VendorProfile.exportPerformance. Request shapes come from ZodValidationPipe schemas (single source). Response DTOs are not yet Zod-exported — SuccessEnvelope is a placeholder.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "nullable": true
            },
            "required": false,
            "name": "from",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "nullable": true
            },
            "required": false,
            "name": "to",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "FIND_ORNAMENT",
                "SELL_OLD_GOLD",
                "GOLD_COIN",
                "GOLD_BULLION"
              ]
            },
            "required": false,
            "name": "requestType",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": false,
            "name": "categoryId",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": false,
            "name": "regionId",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Success (envelope per Architecture §13 / EnvelopeInterceptor)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
  }
}